requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.10.0
plotly>=5.10.0
joblib>=1.1.0
python-dotenv>=0.19.0

# Advanced ML Models
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
prophet>=1.1.0
statsmodels>=0.13.0

# Explainable AI
shap>=0.41.0
lime>=0.2.0

# Optimization
optuna>=3.0.0
onnx>=1.13.0
onnxruntime>=1.13.0

# API & Deployment
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# NLP & Multimodal
transformers>=4.26.0
torch>=2.0.0
Pillow>=9.4.0
openai>=0.27.0

# Anomaly Detection
pyod>=1.0.0